requirements.txt
alembic==0.8.9
antiorm==1.2.1
appnope==0.1.0
click==6.6
decorator==4.0.11
Flask==0.11.1
Flask-Migrate==2.0.2
Flask-Script==2.0.5
Flask-Session==0.3.1
Flask-SQLAlchemy==2.1
future==0.16.0
google-api-python-client==1.5.5
gunicorn==19.6.0
httplib2==0.9.2
ipython==5.1.0
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
line-bot-sdk==1.0.2
Mako==1.0.6
MarkupSafe==0.23
oauth2client==4.0.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.2
ptyprocess==0.5.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
Pygments==2.1.3
python-dotenv==0.6.2
python-editor==1.0.3
PyYAML==3.12
requests==2.12.3
rsa==3.4.2
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.1.4
traitlets==4.3.1
uritemplate==3.0.0
wcwidth==0.1.7
Werkzeug==0.11.11
